Marge into one file all controllers related to one entity (e.g. Group), remove unused get_summary_controller.py, add annotations to controllers
